specular

+++ src/vec3f.c
@@ -2,7 +2,7 @@
 
 float	magnitude(t_vec3f v)
 {
-	return (sqrt(dot_product(v, v))); // sqrtf??
+	return (sqrtf(dot_product(v, v))); // sqrtf??
 }
 
 float	dot_product(t_vec3f u, t_vec3f v)
@@ -20,7 +20,7 @@ t_vec3f	cross_product(t_vec3f u, t_vec3f v)
 	return (cross_product);
 }
 
-t_vec3f	normalize(t_vec3f v) // multiplication more efficient then division
+t_vec3f	normalize(t_vec3f v)
 {
 	float	len;
 	float	inv_len;
@@ -35,54 +35,3 @@ t_vec3f	normalize(t_vec3f v) // multiplication more efficient then division
 	}
 	return (v);
 }
-
-// void	set_values(t_vec3 *vec3, float x, float y, float z)
-// {
-// 	vec3->x = x;
-// 	vec3->y = y;
-// 	vec3->z = z;
-// }
-
-// float	magnitude(t_vec3 v)
-// {
-// 	return (sqrt(dot_product(v, v))); // sqrtf??
-// }
-
-// float	dot_product(t_vec3 u, t_vec3 v)
-// {
-// 	return (u.x * v.x + u.y * v.y + u.z * v.z);
-// }
-
-// t_vec3	cross_product(t_vec3 u, t_vec3 v)
-// {
-// 	t_vec3 vec3;
-
-// 	vec3.x = u.y * v.z - u.z * v.y;
-// 	vec3.y = u.z * v.x - u.x * v.z;
-// 	vec3.z = u.x * v.y - u.y * v.x;
-// 	return (vec3);
-// }
-
-// t_vec3	normalize(t_vec3 v) // multiplication more efficient then division
-// {
-// 	float	len;
-// 	float	inv_len;
-	
-// 	len = magnitude(v);
-// 	if (len > 0)
-// 	{
-// 		inv_len = 1 / len;
-// 		v.x *= inv_len;
-// 		v.y *= inv_len;
-// 		v.z *= inv_len;
-// 	}
-// 	return (v);
-// }
-
-// void	set_values(t_vec3 *vec3, float x, float y, float z)
-// {
-// 	vec3->x = x;
-// 	vec3->y = y;
-// 	vec3->z = z;
-// }
-
